To find all MAPL edits needed for GCHP search for string "GCHP".

# Conflicts:
#	CMakeLists.txt
#	MAPL/CMakeLists.txt
#	MAPL_cfio/CMakeLists.txt
#	base/CMakeLists.txt
#	base/MAPL_ExtDataGridCompMod.F90
#	pfio/CMakeLists.txt
#	shared/CMakeLists.txt
#	shared/MAPL_ErrorHandling.F90

+++ base/MAPL_MemUtils.F90
@@ -572,7 +572,7 @@ integer, optional, intent(OUT  ) :: RC
 character(len=32) :: proc_self = '/proc/self/status'
 character(len=32) :: meminfo   = '/proc/meminfo'
 character(len=32) :: string
-integer :: memtot, memfree, swaptot, swapfree
+integer(kind=INT64) :: memtot, memfree, swaptot, swapfree
 integer :: mem_unit
 real    :: multiplier
